Refactored opening times text into own component
